chore(fluentui/e2e): speedup ts processing and make it work with no-build DX
