+++ packages/fluentui/e2e/tests/carouselClickableContent-selectors.ts
@@ -1,0 +1,7 @@
+export const selectors = {
+  CarouselClass: 'carousel',
+  ItemButton: 'item-button',
+  HiddenContent: 'hidden-content',
+  Navigation: 'ui-carousel__navigationitem',
+  second: 'second',
+};
