+++ packages/fluentui/e2e/tests/dialog-selectors.ts
@@ -1,0 +1,4 @@
+export const selectors = {
+  trigger: 'trigger',
+  cancelButton: 'cancelButton',
+};
