+++ packages/fluentui/e2e/tests/popupDismissScroll-example.tsx
@@ -1,25 +1,7 @@
 import React from 'react';
 import { Button, Popup, Flex } from '@fluentui/react-northstar';
 
-export const selectors = {
-  simplePopup: {
-    triggerId: 'trigger',
-    contentId: 'content',
-  },
-  contextPopup: {
-    triggerId: 'trigger-context',
-    contentId: 'content-context',
-  },
-  dismissScrollPopup: {
-    triggerId: 'trigger-dismiss',
-    contentId: 'content-dismiss',
-  },
-  nestedPopup: {
-    parentPopupTriggerId: 'nested-parent-trigger',
-    childPopupTriggerId: 'nested-child-trigger',
-    childPopupContentId: 'nested-child-content',
-  },
-};
+import { selectors } from './popupDismissScroll-selectors';
 
 const PopupClickHandlingExample = () => {
   return (
