+++ packages/fluentui/e2e/tests/popupDismissScroll-selectors.ts
@@ -1,0 +1,19 @@
+export const selectors = {
+  simplePopup: {
+    triggerId: 'trigger',
+    contentId: 'content',
+  },
+  contextPopup: {
+    triggerId: 'trigger-context',
+    contentId: 'content-context',
+  },
+  dismissScrollPopup: {
+    triggerId: 'trigger-dismiss',
+    contentId: 'content-dismiss',
+  },
+  nestedPopup: {
+    parentPopupTriggerId: 'nested-parent-trigger',
+    childPopupTriggerId: 'nested-child-trigger',
+    childPopupContentId: 'nested-child-content',
+  },
+};
