+++ packages/fluentui/e2e/tests/popupIframeDismiss-selectors.ts
@@ -1,0 +1,5 @@
+export const selectors = {
+  popupTriggerId: 'trigger',
+  popupContentClass: 'ui-popup__content',
+  iframe: 'iframe',
+};
