+++ packages/fluentui/e2e/tests/popupWithTooltipTrigger-selectors.ts
@@ -1,0 +1,4 @@
+export const selectors = {
+  trigger: 'tooltip-trigger',
+  content: 'tooltip-content',
+};
