+++ packages/fluentui/e2e/tests/toolbarWithPopupTooltip-selectors.ts
@@ -1,0 +1,11 @@
+export const clickSelectors = {
+  triggerId: 'trigger-click',
+  contentId: 'content-click',
+  popupId: 'popup-click',
+};
+
+export const hoverSelectors = {
+  triggerId: 'trigger-hover',
+  contentId: 'content-hover',
+  popupId: 'popup-hover',
+};
