+++ packages/fluentui/e2e/tests/tooltipUnhandledProps-selectors.ts
@@ -1,0 +1,4 @@
+export const selectors = {
+  menuItemClassName: 'menu-item',
+  triggerClassName: 'trigger-button',
+};
